Add "notify" action to manually send notification emails for new submission.
