+++ src/main/java/site/controller/CfpController.java
@@ -17,10 +17,6 @@ import site.model.Submission;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
-import java.io.IOException;
-import java.net.URISyntaxException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 
 /**
  * @author Ivan St. Ivanov
@@ -32,8 +28,6 @@ public class CfpController extends AbstractCfpController {
 
     public static final String CFP_OPEN_JSP = "/proposal.jsp";
     public static final String CFP_CLOSED_JSP = "/cfp-closed.jsp";
-
-    public static final String JPRIME_CONF_MAIL_ADDRESS = "conference@jprime.io";
 
     @RequestMapping(value = "/cfp", method = RequestMethod.GET)
     public String submissionForm(Model model) {
@@ -70,15 +64,7 @@ public class CfpController extends AbstractCfpController {
 
         saveSubmission(submission, speakerImage, coSpeakerImage);
         try {
-            mailFacade.sendEmail(submission.getSpeaker().getEmail(), "jPrime talk proposal",
-                    loadMailContentTemplate("submissionContent.html"));
-            if (submission.getCoSpeaker() != null) {
-                mailFacade.sendEmail(submission.getCoSpeaker().getEmail(), "jPrime talk proposal",
-                        loadMailContentTemplate("submissionContent.html"));
-            }
-            mailFacade.sendEmail(JPRIME_CONF_MAIL_ADDRESS, "New talk proposal",
-                    prepareNewSubmissionContent(submission, loadMailContentTemplate("newSubmission.html")
-            ));
+           sendNotificationEmails(submission);
         } catch (Exception e) {
             logger.error("Could not send confirmation email", e);
         }
@@ -102,17 +88,5 @@ public class CfpController extends AbstractCfpController {
         model.addAttribute("tags", userFacade.findAllTags());
         buildCfpFormModel(model, submission);
         return Globals.PAGE_CFP;
-    }
-
-    private String loadMailContentTemplate(String templateFileName)
-            throws IOException, URISyntaxException {
-        return new String(Files.readAllBytes(Paths.get(getClass().getResource("/" + templateFileName).toURI())));
-    }
-
-    private String prepareNewSubmissionContent(Submission submission, String template) {
-        return template.replace("{session.title}", submission.getTitle())
-                        .replace("{session.abstract}", submission.getDescription())
-                        .replace("{speaker.name}", submission.getSpeaker().getFirstName() + " " + submission.getSpeaker().getLastName())
-                        .replace("{speaker.bio}", submission.getSpeaker().getBio());
     }
 }
